# app-code/src/openapi_server/apis/default_api.py
from typing import Dict, List, Any

from fastapi import (
    APIRouter,
    Body,
    Cookie,
    Depends,
    Form,
    Header,
    HTTPException,
    Path,
    Query,
    Request,
    Response,
    Security,
    status,
)

from openapi_server.factories.factory import generic_mock_model
from openapi_server.models.extra_models import TokenModel

from pydantic import StrictInt
from typing import Any, List
from openapi_server.models.driver import Driver
from openapi_server.models.race import Race
from openapi_server.models.result import Result
from openapi_server.models.team import Team

from openapi_server.security_api import get_token_bearerAuth

router = APIRouter()

status_codes = {
    "get": 200,
    "post": 201,
    "put": 200,
    "patch": 200,
    "delete": 200
}

@router.delete(
    "/drivers/{driverId}",
    responses={
        "204": {"description": "Deleted"},
        "401": {"description": "Unauthorized"},
        "404": {"description": "Driver not found"},
        "500": {"description": "Internal server error"},
    },
    tags=["default"],
    summary="Delete a driver by ID",
    response_model_by_alias=True
)
async def drivers_driver_id_delete(
    response: Response,
    driverId: StrictInt = Path(..., description=""),
    token_bearerAuth: TokenModel = Security(
        get_token_bearerAuth
    ),
):
    raise HTTPException(status_code=501, detail="Not implemented")

@router.get(
    "/drivers/{driverId}",
    responses={
        "200": {"model": Driver, "description": "A driver"},
        "401": {"description": "Unauthorized"},
        "404": {"description": "Driver not found"},
        "500": {"description": "Internal server error"},
    },
    tags=["default"],
    summary="Get a driver by ID",
    response_model_by_alias=True
)
async def drivers_driver_id_get(
    response: Response,
    driverId: StrictInt = Path(..., description=""),
    token_bearerAuth: TokenModel = Security(
        get_token_bearerAuth
    ),
):
    if status_codes["get"]:
        response.status_code = status_codes["get"]
    return generic_mock_model(Driver)

@router.put(
    "/drivers/{driverId}",
    responses={
        "200": {"model": Driver, "description": "Updated"},
        "400": {"description": "Bad request"},
        "404": {"description": "Driver not found"},
        "500": {"description": "Internal server error"},
    },
    tags=["default"],
    summary="Update a driver by ID",
    response_model_by_alias=True
)
async def drivers_driver_id_put(
    response: Response,
    driverId: StrictInt = Path(..., description=""),
    driver: Driver = Body(None, description=""),
    token_bearerAuth: TokenModel = Security(
        get_token_bearerAuth
    ),
):
    if status_codes["put"]:
        response.status_code = status_codes["put"]
    return generic_mock_model(Driver)

@router.get(
    "/drivers",
    responses={
        "200": {"model": List[Driver], "description": "A list of drivers"},
        "401": {"description": "Unauthorized"},
        "500": {"description": "Internal server error"},
    },
    tags=["default"],
    summary="Get all drivers",
    response_model_by_alias=True
)
async def drivers_get(
    response: Response,
    token_bearerAuth: TokenModel = Security(
        get_token_bearerAuth
    ),
):
    if status_codes["get"]:
        response.status_code = status_codes["get"]
    return generic_mock_model(List[Driver])

@router.post(
    "/drivers",
    responses={
        "201": {"model": Driver, "description": "Created"},
        "400": {"description": "Bad request"},
        "500": {"description": "Internal server error"},
    },
    tags=["default"],
    summary="Create a new driver",
    response_model_by_alias=True
)
async def drivers_post(
    response: Response,
    driver: Driver = Body(None, description=""),
    token_bearerAuth: TokenModel = Security(
        get_token_bearerAuth
    ),
):
    if status_codes["post"]:
        response.status_code = status_codes["post"]
    return generic_mock_model(Driver)

@router.get(
    "/races",
    responses={
        "200": {"model": List[Race], "description": "A list of races"},
        "401": {"description": "Unauthorized"},
        "500": {"description": "Internal server error"},
    },
    tags=["default"],
    summary="Get all races",
    response_model_by_alias=True
)
async def races_get(
    response: Response,
    token_bearerAuth: TokenModel = Security(
        get_token_bearerAuth
    ),
):
    if status_codes["get"]:
        response.status_code = status_codes["get"]
    return generic_mock_model(List[Race])

@router.post(
    "/races",
    responses={
        "201": {"model": Race, "description": "Created"},
        "400": {"description": "Bad request"},
        "500": {"description": "Internal server error"},
    },
    tags=["default"],
    summary="Create a new race",
    response_model_by_alias=True
)
async def races_post(
    response: Response,
    race: Race = Body(None, description=""),
    token_bearerAuth: TokenModel = Security(
        get_token_bearerAuth
    ),
):
    if status_codes["post"]:
        response.status_code = status_codes["post"]
    return generic_mock_model(Race)

@router.delete(
    "/races/{raceId}",
    responses={
        "204": {"description": "Deleted"},
        "401": {"description": "Unauthorized"},
        "404": {"description": "Race not found"},
        "500": {"description": "Internal server error"},
    },
    tags=["default"],
    summary="Delete a race by ID",
    response_model_by_alias=True
)
async def races_race_id_delete(
    response: Response,
    raceId: StrictInt = Path(..., description=""),
    token_bearerAuth: TokenModel = Security(
        get_token_bearerAuth
    ),
):
    raise HTTPException(status_code=501, detail="Not implemented")

@router.get(
    "/races/{raceId}",
    responses={
        "200": {"model": Race, "description": "A race"},
        "401": {"description": "Unauthorized"},
        "404": {"description": "Race not found"},
        "500": {"description": "Internal server error"},
    },
    tags=["default"],
    summary="Get a race by ID",
    response_model_by_alias=True
)
async def races_race_id_get(
    response: Response,
    raceId: StrictInt = Path(..., description=""),
    token_bearerAuth: TokenModel = Security(
        get_token_bearerAuth
    ),
):
    if status_codes["get"]:
        response.status_code = status_codes["get"]
    return generic_mock_model(Race)

@router.put(
    "/races/{raceId}",
    responses={
        "200": {"model": Race, "description": "Updated"},
        "400": {"description": "Bad request"},
        "404": {"description": "Race not found"},
        "500": {"description": "Internal server error"},
    },
    tags=["default"],
    summary="Update a race by ID",
    response_model_by_alias=True
)
async def races_race_id_put(
    response: Response,
    raceId: StrictInt = Path(..., description=""),
    race: Race = Body(None, description=""),
    token_bearerAuth: TokenModel = Security(
        get_token_bearerAuth
    ),
):
    if status_codes["put"]:
        response.status_code = status_codes["put"]
    return generic_mock_model(Race)

@router.get(
    "/races/{raceId}/results",
    responses={
        "200": {"model": List[Result], "description": "A list of results for the race"},
        "401": {"description": "Unauthorized"},
        "404": {"description": "Race not found"},
        "500": {"description": "Internal server error"},
    },
    tags=["default"],
    summary="Get results for a race by ID",
    response_model_by_alias=True
)
async def races_race_id_results_get(
    response: Response,
    raceId: StrictInt = Path(..., description=""),
    token_bearerAuth: TokenModel = Security(
        get_token_bearerAuth
    ),
):
    if status_codes["get"]:
        response.status_code = status_codes["get"]
    return generic_mock_model(List[Result])

@router.post(
    "/races/{raceId}/results",
    responses={
        "201": {"model": Result, "description": "Created"},
        "400": {"description": "Bad request"},
        "404": {"description": "Race not found"},
        "500": {"description": "Internal server error"},
    },
    tags=["default"],
    summary="Create a new result for a race",
    response_model_by_alias=True
)
async def races_race_id_results_post(
    response: Response,
    raceId: StrictInt = Path(..., description=""),
    result: Result = Body(None, description=""),
    token_bearerAuth: TokenModel = Security(
        get_token_bearerAuth
    ),
):
    if status_codes["post"]:
        response.status_code = status_codes["post"]
    return generic_mock_model(Result)

@router.get(
    "/teams",
    responses={
        "200": {"model": List[Team], "description": "A list of teams"},
        "401": {"description": "Unauthorized"},
        "500": {"description": "Internal server error"},
    },
    tags=["default"],
    summary="Get all teams",
    response_model_by_alias=True
)
async def teams_get(
    response: Response,
    token_bearerAuth: TokenModel = Security(
        get_token_bearerAuth
    ),
):
    if status_codes["get"]:
        response.status_code = status_codes["get"]
    return generic_mock_model(List[Team])

@router.post(
    "/teams",
    responses={
        "201": {"model": Team, "description": "Created"},
        "400": {"description": "Bad request"},
        "500": {"description": "Internal server error"},
    },
    tags=["default"],
    summary="Create a new team",
    response_model_by_alias=True
)
async def teams_post(
    response: Response,
    team: Team = Body(None, description=""),
    token_bearerAuth: TokenModel = Security(
        get_token_bearerAuth
    ),
):
    if status_codes["post"]:
        response.status_code = status_codes["post"]
    return generic_mock_model(Team)

@router.delete(
    "/teams/{teamId}",
    responses={
        "204": {"description": "Deleted"},
        "401": {"description": "Unauthorized"},
        "404": {"description": "Team not found"},
        "500": {"description": "Internal server error"},
    },
    tags=["default"],
    summary="Delete a team by ID",
    response_model_by_alias=True
)
async def teams_team_id_delete(
    response: Response,
    teamId: StrictInt = Path(..., description=""),
    token_bearerAuth: TokenModel = Security(
        get_token_bearerAuth
    ),
):
    raise HTTPException(status_code=501, detail="Not implemented")

@router.get(
    "/teams/{teamId}",
    responses={
        "200": {"model": Team, "description": "A team"},
        "401": {"description": "Unauthorized"},
        "404": {"description": "Team not found"},
        "500": {"description": "Internal server error"},
    },
    tags=["default"],
    summary="Get a team by ID",
    response_model_by_alias=True
)
async def teams_team_id_get(
    response: Response,
    teamId: StrictInt = Path(..., description=""),
    token_bearerAuth: TokenModel = Security(
        get_token_bearerAuth
    ),
):
    if status_codes["get"]:
        response.status_code = status_codes["get"]
    return generic_mock_model(Team)

@router.put(
    "/teams/{teamId}",
    responses={
        "200": {"model": Team, "description": "Updated"},
        "400": {"description": "Bad request"},
        "404": {"description": "Team not found"},
        "500": {"description": "Internal server error"},
    },
    tags=["default"],
    summary="Update a team by ID",
    response_model_by_alias=True
)
async def teams_team_id_put(
    response: Response,
    teamId: StrictInt = Path(..., description=""),
    team: Team = Body(None, description=""),
    token_bearerAuth: TokenModel = Security(
        get_token_bearerAuth
    ),
):
    if status_codes["put"]:
        response.status_code = status_codes["put"]
    return generic_mock_model(Team)